+++ packages/extensions/lib/src/logging/logging_builder_configuration_extensions.dart
@@ -4,9 +4,9 @@ import 'i_logger_provider_configuration_factory.dart';
 import 'logger_provider_configuration_factory.dart';
 import 'logging_builder.dart';
 
-/// Extension methods for configuring logging from [IConfiguration].
+/// Extension methods for configuring logging from [Configuration].
 extension LoggingBuilderConfigurationExtensions on LoggingBuilder {
-  /// Adds services required to consume [ILoggerProviderConfigurationFactory].
+  /// Adds services required to consume [LoggerProviderConfigurationFactory].
   ///
   /// This method registers the configuration factory that allows logger
   /// providers to retrieve their specific configuration from the application's
@@ -20,8 +20,8 @@ extension LoggingBuilderConfigurationExtensions on LoggingBuilder {
   /// ```
   ///
   /// After calling this method, logger providers can inject
-  /// [ILoggerProviderConfigurationFactory] to access their configuration.
-  LoggingBuilder addConfiguration(IConfiguration configuration) {
+  /// [LoggerProviderConfigurationFactory] to access their configuration.
+  LoggingBuilder addConfiguration(Configuration configuration) {
     // Register the configuration factory as a singleton
     services.tryAddSingleton<LoggerProviderConfigurationFactory>(
       (sp) => LoggerProviderConfigurationFactoryImpl([configuration]),
@@ -30,7 +30,7 @@ extension LoggingBuilderConfigurationExtensions on LoggingBuilder {
     return this;
   }
 
-  /// Adds services required to consume [ILoggerProviderConfigurationFactory]
+  /// Adds services required to consume [LoggerProviderConfigurationFactory]
   /// from multiple configuration sources.
   ///
   /// This overload allows you to provide multiple configuration sources that
@@ -46,7 +46,7 @@ extension LoggingBuilderConfigurationExtensions on LoggingBuilder {
   ///   ]);
   /// });
   /// ```
-  LoggingBuilder addConfigurations(Iterable<IConfiguration> configurations) {
+  LoggingBuilder addConfigurations(Iterable<Configuration> configurations) {
     // Register the configuration factory with multiple sources
     services.tryAddSingleton<LoggerProviderConfigurationFactory>(
       (sp) => LoggerProviderConfigurationFactoryImpl(configurations),